Self-review: Cleanup code and comments
